+++ app/src/main/res/drawable/ic_movies.xml
@@ -1,0 +1,5 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:viewportHeight="24" android:viewportWidth="24" android:width="24dp">
+      
+    <path android:fillColor="#FF000000" android:pathData="m22.772,11.125c-0.647,-0.256 -1.387,-0.102 -1.945,0.314l-0.982,0.789c-0.33,-1.384 -1.18,-2.566 -2.336,-3.326 0.919,-0.946 1.491,-2.232 1.491,-3.652 0,-2.895 -2.355,-5.25 -5.25,-5.25 -1.47,0 -2.796,0.611 -3.75,1.587 -0.954,-0.976 -2.28,-1.587 -3.75,-1.587C3.355,0 1,2.355 1,5.25c0,1.42 0.572,2.706 1.491,3.652 -1.498,0.984 -2.491,2.676 -2.491,4.598v5c0,3.033 2.467,5.5 5.5,5.5h9c2.565,0 4.72,-1.768 5.325,-4.148l1.088,0.763c0.342,0.255 0.746,0.385 1.153,0.385 0.394,0 0.791,-0.122 1.137,-0.369 0.525,-0.374 0.797,-1.015 0.797,-1.66v-6.047c0,-0.794 -0.476,-1.502 -1.228,-1.8ZM13.75,3c1.241,0 2.25,1.009 2.25,2.25s-1.009,2.25 -2.25,2.25 -2.25,-1.009 -2.25,-2.25 1.009,-2.25 2.25,-2.25ZM6.25,3c1.241,0 2.25,1.009 2.25,2.25s-1.009,2.25 -2.25,2.25 -2.25,-1.009 -2.25,-2.25 1.009,-2.25 2.25,-2.25ZM17,18.5c0,1.378 -1.122,2.5 -2.5,2.5L5.5,21c-1.378,0 -2.5,-1.122 -2.5,-2.5v-5c0,-1.378 1.122,-2.5 2.5,-2.5h9c1.378,0 2.5,1.122 2.5,2.5v5Z"/>
+    
+</vector>
